typecheck_struct_ctor
